Attack!

+++ src/main/java/com/zuehlke/securesoftwaredevelopment/controller/PersonsController.java
@@ -1,6 +1,7 @@
 package com.zuehlke.securesoftwaredevelopment.controller;
 
 import com.zuehlke.securesoftwaredevelopment.config.AuditLogger;
+import com.zuehlke.securesoftwaredevelopment.config.SecurityUtil;
 import com.zuehlke.securesoftwaredevelopment.domain.Person;
 import com.zuehlke.securesoftwaredevelopment.domain.User;
 import com.zuehlke.securesoftwaredevelopment.repository.PersonRepository;
@@ -8,11 +9,16 @@ import com.zuehlke.securesoftwaredevelopment.repository.UserRepository;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
+import java.nio.file.AccessDeniedException;
 import java.sql.SQLException;
 import java.util.List;
 
@@ -32,20 +38,32 @@ public class PersonsController {
     }
 
     @GetMapping("/persons/{id}")
-    public String person(@PathVariable int id, Model model) {
+    @PreAuthorize("hasAuthority('VIEW_PERSON')")
+    public String person(@PathVariable int id, Model model, HttpSession session) {
+        String csrf = session.getAttribute("CSRF_TOKEN").toString();
+        model.addAttribute("CSRF_TOKEN", csrf);
         model.addAttribute("person", personRepository.get("" + id));
         return "person";
     }
 
     @GetMapping("/myprofile")
-    public String self(Model model, Authentication authentication) {
+    @PreAuthorize("hasAuthority('VIEW_MY_PROFILE')")
+    public String self(Model model, Authentication authentication, HttpSession session) {
         User user = (User) authentication.getPrincipal();
+        String csrf = session.getAttribute("CSRF_TOKEN").toString();
+        model.addAttribute("CSRF_TOKEN", csrf);
         model.addAttribute("person", personRepository.get("" + user.getId()));
         return "person";
     }
 
     @DeleteMapping("/persons/{id}")
-    public ResponseEntity<Void> person(@PathVariable int id) {
+    @PreAuthorize("hasAuthority('VIEW_PERSON')")
+    public ResponseEntity<Void> person(@PathVariable int id, HttpSession session, @RequestParam("csrfToken") String csrfToken) throws  AccessDeniedException{
+        String sessionToken = session.getAttribute("CSRF_TOKEN").toString();
+        if (!csrfToken.equals(sessionToken))
+        {
+            throw new AccessDeniedException("Access forbidden! " + csrfToken + "!=" + sessionToken);
+        }
         personRepository.delete(id);
         userRepository.delete(id);
 
@@ -53,12 +71,31 @@ public class PersonsController {
     }
 
     @PostMapping("/update-person")
-    public String updatePerson(Person person) {
-        personRepository.update(person);
-        return "redirect:/persons/" + person.getId();
+    @PreAuthorize("hasAuthority('UPDATE_PERSON')")
+    public String updatePerson(Person person, HttpSession session, @RequestParam("csrfToken") String csrfToken)
+            throws AccessDeniedException {
+        String sessionToken = session.getAttribute("CSRF_TOKEN").toString();
+        if (!csrfToken.equals(sessionToken))
+        {
+            throw new AccessDeniedException("Access forbidden! " + csrfToken + "!=" + sessionToken);
+        }
+
+        if (SecurityUtil.hasPermission("VIEW_PERSON")){
+            personRepository.update(person);
+            return "redirect:/persons/" + person.getId();
+        }
+        else {
+            int userID = SecurityUtil.getCurrentUser().getId();
+            if (userID != Integer.parseInt(person.getId())){
+                throw new AccessDeniedException("Access forbidden!");
+            }
+            personRepository.update(person);
+            return "redirect:/myprofile";
+        }
     }
 
     @GetMapping("/persons")
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     public String persons(Model model) {
         model.addAttribute("persons", personRepository.getAll());
         return "persons";
@@ -66,6 +103,7 @@ public class PersonsController {
 
     @GetMapping(value = "/persons/search", produces = "application/json")
     @ResponseBody
+    @PreAuthorize("hasAuthority('VIEW_PERSONS_LIST')")
     public List<Person> searchPersons(@RequestParam String searchTerm) throws SQLException {
         return personRepository.search(searchTerm);
     }
